serialization refactor file structure
